chore: rename version command to services
